Move circuit constructors to shared.
