+++ cpp/src/barretenberg/honk/composer/composer_helper/turbo_plonk_composer_helper.cpp
@@ -1,5 +1,5 @@
 #include "turbo_plonk_composer_helper.hpp"
-#include "barretenberg/honk/circuit_constructors/turbo_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_constructors/turbo_circuit_constructor.hpp"
 #include "barretenberg/ecc/curves/bn254/scalar_multiplication/scalar_multiplication.hpp"
 #include "barretenberg/numeric/bitop/get_msb.hpp"
 #include "barretenberg/plonk/proof_system/widgets/random_widgets/permutation_widget.hpp"
