+++ cpp/src/barretenberg/honk/composer/standard_plonk_composer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "composer_helper/standard_plonk_composer_helper.hpp"
-#include "barretenberg/honk/circuit_constructors/standard_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_constructors/standard_circuit_constructor.hpp"
 #include "barretenberg/srs/reference_string/file_reference_string.hpp"
 #include "barretenberg/transcript/manifest.hpp"
 #include "barretenberg/proof_system/flavor/flavor.hpp"
