+++ cpp/src/barretenberg/proof_system/circuit_constructors/circuit_constructor_base.cpp
